Fixing issue - https://github.com/ballerinalang/composer/issues/1072 and fixing the source generation issue for functions.

+++ modules/web/js/ballerina/views/input-struct-view.js
@@ -15,8 +15,8 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-define(['lodash', 'log','./ballerina-view','./../ast/resource-parameter', 'typeMapper','constants'],
-    function (_, log, BallerinaView,ResourceParameter, TypeMapperRenderer,Constants) {
+define(['lodash', 'log', './ballerina-view', './../ast/resource-parameter', 'typeMapper', 'constants'],
+    function (_, log, BallerinaView, ResourceParameter, TypeMapperRenderer, Constants) {
 
         var InputStructView = function (args) {
             BallerinaView.call(this, args);
@@ -48,16 +48,17 @@ define(['lodash', 'log','./ballerina-view','./../ast/resource-parameter', 'typeM
             this._diagramRenderingContext = diagramRenderingContext;
             var typeStructSchema = this.getSourceInfo().sourceStruct;
             var previousSelection = this.getSourceInfo()[TYPE_MAPPER_COMBOBOX_PREVIOUS_SELECTION];
+            if (typeStructSchema) {
+                if (!mapper) {
+                    mapper = new TypeMapperRenderer(self.getOnConnectInstance(), self.getOnDisconnectInstance(), this._parentView);
+                    this._parentView._typeMapper = mapper;
+                }
 
-            if(!mapper) {
-                mapper = new TypeMapperRenderer(self.getOnConnectInstance(), self.getOnDisconnectInstance(), this._parentView);
-                this._parentView._typeMapper = mapper;
+                if (previousSelection != undefined && previousSelection != TYPE_MAPPER_COMBOBOX_DEFAULT_SELECTION) {
+                    mapper.removeStruct(previousSelection);
+                }
+                mapper.addSourceStruct(typeStructSchema.getAttributesArray(), this.getModel());
             }
-
-            if(previousSelection != undefined && previousSelection != TYPE_MAPPER_COMBOBOX_DEFAULT_SELECTION){
-                mapper.removeStruct(previousSelection);
-            }
-            mapper.addSourceStruct(typeStructSchema.getAttributesArray(),this.getModel());
         };
 
         /**
@@ -111,4 +112,4 @@ define(['lodash', 'log','./ballerina-view','./../ast/resource-parameter', 'typeM
         };
 
         return InputStructView;
-});
+    });
