+++ modules/web/js/ballerina/views/output-struct-view.js
@@ -15,8 +15,8 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-define(['lodash', 'log','./ballerina-view','./../ast/return-type', 'typeMapper','constants'],
-    function (_, log, BallerinaView,ReturnType, TypeMapperRenderer,Constants) {
+define(['lodash', 'log', './ballerina-view', './../ast/return-type', 'typeMapper', 'constants'],
+    function (_, log, BallerinaView, ReturnType, TypeMapperRenderer, Constants) {
 
         var OutputStructView = function (args) {
             BallerinaView.call(this, args);
@@ -49,16 +49,17 @@ define(['lodash', 'log','./ballerina-view','./../ast/return-type', 'typeMapper',
 
             var typeStructSchema = this.getTargetInfo().targetStruct;
             var previousSelection = this.getTargetInfo()[TYPE_MAPPER_COMBOBOX_PREVIOUS_SELECTION];
+            if (typeStructSchema) {
+                if (!mapper) {
+                    mapper = new TypeMapperRenderer(self.getOnConnectInstance(), self.getOnDisconnectInstance(), this._parentView);
+                    this._parentView._typeMapper = mapper;
+                }
 
-            if(!mapper) {
-                mapper = new TypeMapperRenderer(self.getOnConnectInstance(), self.getOnDisconnectInstance(), this._parentView);
-                this._parentView._typeMapper = mapper;
+                if (previousSelection != undefined && previousSelection != TYPE_MAPPER_COMBOBOX_DEFAULT_SELECTION) {
+                    mapper.removeStruct(previousSelection);
+                }
+                mapper.addTargetStruct(typeStructSchema.getAttributesArray(), this.getModel());
             }
-
-            if(previousSelection != undefined && previousSelection != TYPE_MAPPER_COMBOBOX_DEFAULT_SELECTION){
-                mapper.removeStruct(previousSelection);
-            }
-            mapper.addTargetStruct(typeStructSchema.getAttributesArray(),this.getModel());
         };
 
         /**
@@ -112,4 +113,4 @@ define(['lodash', 'log','./ballerina-view','./../ast/return-type', 'typeMapper',
         };
 
         return OutputStructView;
-});
+    });
